requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
openai==1.12.0
flask-sqlalchemy==3.1.1
Werkzeug==3.0.1
